Finished implementing file size check.
git-svn-id: https://subversion.assembla.com/svn/agon/ArtSizeReader@235 b7dd9bf6-892c-4cbd-8c55-1525a1e9bee7

+++ ArtSizeReader/IArtReader.cs
@@ -16,6 +16,6 @@ namespace ArtSizeReader {
 
         IArtReader WithRatio(bool hasRatio);
 
-        IArtReader WithSize(int size);
+        IArtReader WithSize(double size);
     }
 }
